feat: download assets from public repository's GitHub Release without GitHub API
I have tested if downloading assets from public repository's GitHub Releases anonymously is rate limited.
As a result of test, it seems not to be limited.
So at first aqua tries to download assets without GitHub API.
And if it failed, aqua tries again with GitHub API.
It avoids the rate limit of GitHub API.