fix: improve handling of broken registry JSON files
If it failed to read a registry JSON file, aqua failed.
In that case, users have to remove the file themselves.
This commit improves the behavior so that aqua can handle broken regsistry JSON files well.
If it fails to read a registry JSON file, aqua removes the file, reads a registry YAML file, and recreates a registry JSON file, and continues the process.