fix: improve the completion settings suggested in `aqua completion --help` (#3297)
* fix: suggest ".bashrc" and ".zshrc" for completion setting

The current help text suggests adding the completion settings in
.bash_profile and .zprofile, but .bash_profile and .zprofile are
loaded only in login shell sessions. They are not loaded in
(non-login) interactive sessions started inside the login shells, and
thus the completion would be disabled in non-login interactive
sessions.

In general, .bash_profile and .zprofile should only contain the
settings that are automatically inherited by the child processes or
the global user settings in the operating system.  The interactive
settings should be put in .bashrc and .zshrc.

In this patch, we suggest .bashrc and .zshrc instead of .bash_profile
and .zprofile, respectively.

* fix: fix the example of a completion setting for Bash 3

The source builtin in Bash 3.* does not support reading from a pipe,
so `source <(command)` fails silently.  `eval "$(command)"` works in
all versions of Bash.

* style: make the boundaries of shells outstanding