perf(which): Skip packages that cannot provide the desired exe
When looking for the package providing an exe, several time-consuming
tasks need to be performed on each pkgInfo instance (in particular,
evaluating the version constraints). But in most cases, the exe in
question is not available across *any* version constraints or the like,
in which case we can save some performance by skipping it entirely.

In the case where many packages have to be checked, this can give a
sizable perf boost. Here I'm using `jq`, which is **not** installed by
aqua (and thus every local *and* global package) has to be checked:

```
$ hyperfine --warmup=20 -N 'dist/aqua'{.old,}' which jq'
Benchmark 1: dist/aqua.old which jq
  Time (mean ± σ):       6.1 ms ±   0.7 ms    [User: 5.3 ms, System: 2.1 ms]
  Range (min … max):     4.6 ms …   9.8 ms    573 runs

Benchmark 2: dist/aqua which jq
  Time (mean ± σ):       4.8 ms ±   0.7 ms    [User: 3.9 ms, System: 1.8 ms]
  Range (min … max):     3.3 ms …   9.0 ms    610 runs

Summary
  dist/aqua which jq ran
    1.27 ± 0.24 times faster than dist/aqua.old which jq

$ hyperfine --warmup=20 -N 'dist/aqua'{.old,}' exec -- jq'
Benchmark 1: dist/aqua.old exec -- jq
  Time (mean ± σ):       7.8 ms ±   0.9 ms    [User: 6.6 ms, System: 2.5 ms]
  Range (min … max):     5.8 ms …  11.4 ms    433 runs

Benchmark 2: dist/aqua exec -- jq
  Time (mean ± σ):       6.7 ms ±   0.9 ms    [User: 5.5 ms, System: 2.3 ms]
  Range (min … max):     4.6 ms …  10.4 ms    439 runs

Summary
  dist/aqua exec -- jq ran
    1.18 ± 0.21 times faster than dist/aqua.old exec -- jq
```

If the exe was one that would be found near the start of the list of
packages, this doesn't change very much, but it's at least not a
regression:

```
$ hyperfine --warmup=20 -N 'dist/aqua'{.old,}' which cmdx'
Benchmark 1: dist/aqua.old which cmdx
  Time (mean ± σ):       5.2 ms ±   0.7 ms    [User: 4.4 ms, System: 1.8 ms]
  Range (min … max):     3.5 ms …   8.4 ms    656 runs

Benchmark 2: dist/aqua which cmdx
  Time (mean ± σ):       5.1 ms ±   0.7 ms    [User: 4.3 ms, System: 1.9 ms]
  Range (min … max):     3.6 ms …   8.4 ms    621 runs

Summary
  dist/aqua which cmdx ran
    1.01 ± 0.19 times faster than dist/aqua.old which cmdx

$ hyperfine --warmup=20 -N 'dist/aqua'{.old,}' exec -- cmdx'
Benchmark 1: dist/aqua.old exec -- cmdx
  Time (mean ± σ):       9.3 ms ±   0.9 ms    [User: 7.9 ms, System: 2.9 ms]
  Range (min … max):     7.5 ms …  15.4 ms    373 runs

Benchmark 2: dist/aqua exec -- cmdx
  Time (mean ± σ):       9.2 ms ±   1.2 ms    [User: 7.5 ms, System: 3.1 ms]
  Range (min … max):     7.2 ms …  20.3 ms    338 runs

  Warning: Statistical outliers were detected. Consider re-running this benchmark on a quiet system without any interferences from other programs. It might help to use the '--warmup' or '--prepare' options.

Summary
  dist/aqua exec -- cmdx ran
    1.01 ± 0.17 times faster than dist/aqua.old exec -- cmdx
```